upgrade dyno queues and jedis version
