+++ redis-persistence/src/main/java/com/netflix/conductor/dyno/DynoProxyDiscoveryProvider.java
@@ -1,3 +1,15 @@
+/*
+ * Copyright 2019 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 package com.netflix.conductor.dyno;
 
 import com.netflix.discovery.DiscoveryClient;
@@ -6,7 +18,7 @@ import com.netflix.dyno.jedis.DynoJedisClient;
 import javax.inject.Inject;
 import javax.inject.Provider;
 
-import redis.clients.jedis.JedisCommands;
+import redis.clients.jedis.commands.JedisCommands;
 
 public class DynoProxyDiscoveryProvider implements Provider<JedisCommands> {
     private final DiscoveryClient discoveryClient;
