+++ redis-persistence/src/main/java/com/netflix/conductor/jedis/InMemoryJedisProvider.java
@@ -1,9 +1,20 @@
+/*
+ * Copyright 2019 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 package com.netflix.conductor.jedis;
-
-import redis.clients.jedis.JedisCommands;
 
 import javax.inject.Provider;
 import javax.inject.Singleton;
+import redis.clients.jedis.commands.JedisCommands;
 
 @Singleton
 public class InMemoryJedisProvider implements Provider<JedisCommands> {
