+++ redis-persistence/src/main/java/com/netflix/conductor/jedis/JedisMock.java
@@ -1,5 +1,5 @@
-/**
- * Copyright 2016 Netflix, Inc.
+/*
+ * Copyright 2019 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,14 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * 
- */
 package com.netflix.conductor.jedis;
-
-import org.rarefiedredis.redis.IRedisClient;
-import org.rarefiedredis.redis.IRedisSortedSet.ZsetPair;
-import org.rarefiedredis.redis.RedisMock;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -30,13 +23,15 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
 import java.util.stream.Collectors;
-
+import org.rarefiedredis.redis.IRedisClient;
+import org.rarefiedredis.redis.IRedisSortedSet.ZsetPair;
+import org.rarefiedredis.redis.RedisMock;
 import redis.clients.jedis.Jedis;
 import redis.clients.jedis.ScanParams;
 import redis.clients.jedis.ScanResult;
 import redis.clients.jedis.Tuple;
 import redis.clients.jedis.exceptions.JedisException;
-import redis.clients.jedis.params.sortedset.ZAddParams;
+import redis.clients.jedis.params.ZAddParams;
 
 /**
  * @author Viren
@@ -62,17 +57,6 @@ public class JedisMock extends Jedis {
     @Override public String set(final String key, String value) {
         try { 
             return redis.set(key, value);
-        }
-        catch (Exception e) {
-            throw new JedisException(e);
-        }
-    }
-
-
-    @Override public String set(final String key, final String value, final String nxxx, final String expx,
-                                final long time) {
-        try {
-            return redis.set(key, value, nxxx, expx, String.valueOf(time));
         }
         catch (Exception e) {
             throw new JedisException(e);
@@ -744,7 +728,7 @@ public class JedisMock extends Jedis {
 		
 		try {
 			
-			if(params.contains("xx")) {
+			if(params.getParam("xx") != null) {
 				Double existing = redis.zscore(key, member);
 				if(existing == null) {
 					return 0L;
@@ -886,81 +870,7 @@ public class JedisMock extends Jedis {
             throw new JedisException(e);
         }
     }
-    /*
-    public List<String> sort(final String key) {
-        checkIsInMulti();
-        client.sort(key);
-        return client.getMultiBulkReply();
-    }
 
-    public List<String> sort(final String key, final SortingParams sortingParameters) {
-        checkIsInMulti();
-        client.sort(key, sortingParameters);
-        return client.getMultiBulkReply();
-    }
-
-    public List<String> blpop(final int timeout, final String... keys) {
-        return blpop(getArgsAddTimeout(timeout, keys));
-    }
-
-    private String[] getArgsAddTimeout(int timeout, String[] keys) {
-        final int keyCount = keys.length;
-        final String[] args = new String[keyCount + 1];
-        for (int at = 0; at != keyCount; ++at) {
-            args[at] = keys[at];
-        }
-
-        args[keyCount] = String.valueOf(timeout);
-        return args;
-    }
-
-    public List<String> blpop(String... args) {
-        checkIsInMulti();
-        client.blpop(args);
-        client.setTimeoutInfinite();
-        try {
-            return client.getMultiBulkReply();
-        } finally {
-            client.rollbackTimeout();
-        }
-    }
-
-    public List<String> brpop(String... args) {
-        checkIsInMulti();
-        client.brpop(args);
-        client.setTimeoutInfinite();
-        try {
-            return client.getMultiBulkReply();
-        } finally {
-            client.rollbackTimeout();
-        }
-    }
-
-    @Deprecated
-    public List<String> blpop(String arg) {
-        return blpop(new String[] { arg });
-    }
-
-    public List<String> brpop(String arg) {
-        return brpop(new String[] { arg });
-    }
-
-    public Long sort(final String key, final SortingParams sortingParameters, final String dstkey) {
-        checkIsInMulti();
-        client.sort(key, sortingParameters, dstkey);
-        return client.getIntegerReply();
-    }
-
-    public Long sort(final String key, final String dstkey) {
-        checkIsInMulti();
-        client.sort(key, dstkey);
-        return client.getIntegerReply();
-    }
-
-    public List<String> brpop(final int timeout, final String... keys) {
-        return brpop(getArgsAddTimeout(timeout, keys));
-    }
-    */
     @Override public Long zcount(final String key, final double min, final double max) {
         try {
             return redis.zcount(key, min, max);
@@ -1186,7 +1096,7 @@ public class JedisMock extends Jedis {
             Map<String, String> results = mockr.results;
             List<Entry<String, String>> list = results.entrySet().stream().collect(Collectors.toList());
 			ScanResult<Entry<String, String>> result = new ScanResult<Entry<String, String>>("0", list);
-            
+
             return result;
         }
         catch (Exception e) {
@@ -1207,727 +1117,4 @@ public class JedisMock extends Jedis {
             throw new JedisException(e);
         }
     }
-    
-    
-    /*
-    public Long zunionstore(final String dstkey, final ZParams params, final String... sets) {
-        checkIsInMulti();
-        client.zunionstore(dstkey, params, sets);
-        return client.getIntegerReply();
-    }
-
-    public Long zinterstore(final String dstkey, final String... sets) {
-        checkIsInMulti();
-        client.zinterstore(dstkey, sets);
-        return client.getIntegerReply();
-    }
-
-    public Long zinterstore(final String dstkey, final ZParams params, final String... sets) {
-        checkIsInMulti();
-        client.zinterstore(dstkey, params, sets);
-        return client.getIntegerReply();
-    }
-
-    @Override
-    public Long zlexcount(final String key, final String min, final String max) {
-        checkIsInMulti();
-        client.zlexcount(key, min, max);
-        return client.getIntegerReply();
-    }
-
-    @Override
-    public Set<String> zrangeByLex(final String key, final String min, final String max) {
-        checkIsInMulti();
-        client.zrangeByLex(key, min, max);
-        final List<String> members = client.getMultiBulkReply();
-        if (members == null) {
-            return null;
-        }
-        return new LinkedHashSet<String>(members);
-    }
-
-    @Override
-    public Set<String> zrangeByLex(final String key, final String min, final String max,
-                                   final int offset, final int count) {
-        checkIsInMulti();
-        client.zrangeByLex(key, min, max, offset, count);
-        final List<String> members = client.getMultiBulkReply();
-        if (members == null) {
-            return null;
-        }
-        return new LinkedHashSet<String>(members);
-    }
-
-    @Override
-    public Set<String> zrevrangeByLex(String key, String max, String min) {
-        checkIsInMulti();
-        client.zrevrangeByLex(key, max, min);
-        final List<String> members = client.getMultiBulkReply();
-        if (members == null) {
-            return null;
-        }
-        return new LinkedHashSet<String>(members);
-    }
-
-    @Override
-    public Set<String> zrevrangeByLex(String key, String max, String min, int offset, int count) {
-        checkIsInMulti();
-        client.zrevrangeByLex(key, max, min, offset, count);
-        final List<String> members = client.getMultiBulkReply();
-        if (members == null) {
-            return null;
-        }
-        return new LinkedHashSet<String>(members);
-    }
-
-    @Override
-    public Long zremrangeByLex(final String key, final String min, final String max) {
-        checkIsInMulti();
-        client.zremrangeByLex(key, min, max);
-        return client.getIntegerReply();
-    }
-
-    public Long strlen(final String key) {
-        client.strlen(key);
-        return client.getIntegerReply();
-    }
-
-    public Long lpushx(final String key, final String... string) {
-        client.lpushx(key, string);
-        return client.getIntegerReply();
-    }
-
-    public Long persist(final String key) {
-        client.persist(key);
-        return client.getIntegerReply();
-    }
-
-    public Long rpushx(final String key, final String... string) {
-        client.rpushx(key, string);
-        return client.getIntegerReply();
-    }
-
-    public String echo(final String string) {
-        client.echo(string);
-        return client.getBulkReply();
-    }
-
-    public Long linsert(final String key, final LIST_POSITION where, final String pivot,
-                        final String value) {
-        client.linsert(key, where, pivot, value);
-        return client.getIntegerReply();
-    }
-
-    public String brpoplpush(String source, String destination, int timeout) {
-        client.brpoplpush(source, destination, timeout);
-        client.setTimeoutInfinite();
-        try {
-            return client.getBulkReply();
-        } finally {
-            client.rollbackTimeout();
-        }
-    }
-
-    public Boolean setbit(String key, long offset, boolean value) {
-        client.setbit(key, offset, value);
-        return client.getIntegerReply() == 1;
-    }
-
-    public Boolean setbit(String key, long offset, String value) {
-        client.setbit(key, offset, value);
-        return client.getIntegerReply() == 1;
-    }
-
-    public Boolean getbit(String key, long offset) {
-        client.getbit(key, offset);
-        return client.getIntegerReply() == 1;
-    }
-
-    public Long setrange(String key, long offset, String value) {
-        client.setrange(key, offset, value);
-        return client.getIntegerReply();
-    }
-
-    public String getrange(String key, long startOffset, long endOffset) {
-        client.getrange(key, startOffset, endOffset);
-        return client.getBulkReply();
-    }
-
-    public Long bitpos(final String key, final boolean value) {
-        return bitpos(key, value, new BitPosParams());
-    }
-
-    public Long bitpos(final String key, final boolean value, final BitPosParams params) {
-        client.bitpos(key, value, params);
-        return client.getIntegerReply();
-    }
-
-    public List<String> configGet(final String pattern) {
-        client.configGet(pattern);
-        return client.getMultiBulkReply();
-    }
-
-    public String configSet(final String parameter, final String value) {
-        client.configSet(parameter, value);
-        return client.getStatusCodeReply();
-    }
-
-    public Object eval(String script, int keyCount, String... params) {
-        client.setTimeoutInfinite();
-        try {
-            client.eval(script, keyCount, params);
-            return getEvalResult();
-        } finally {
-            client.rollbackTimeout();
-        }
-    }
-
-    public void subscribe(final JedisPubSub jedisPubSub, final String... channels) {
-        client.setTimeoutInfinite();
-        try {
-            jedisPubSub.proceed(client, channels);
-        } finally {
-            client.rollbackTimeout();
-        }
-    }
-
-    public Long publish(final String channel, final String message) {
-        checkIsInMulti();
-        connect();
-        client.publish(channel, message);
-        return client.getIntegerReply();
-    }
-
-    public void psubscribe(final JedisPubSub jedisPubSub, final String... patterns) {
-        checkIsInMulti();
-        client.setTimeoutInfinite();
-        try {
-            jedisPubSub.proceedWithPatterns(client, patterns);
-        } finally {
-            client.rollbackTimeout();
-        }
-    }
-
-    protected static String[] getParams(List<String> keys, List<String> args) {
-        int keyCount = keys.size();
-        int argCount = args.size();
-
-        String[] params = new String[keyCount + args.size()];
-
-        for (int i = 0; i < keyCount; i++)
-            params[i] = keys.get(i);
-
-        for (int i = 0; i < argCount; i++)
-            params[keyCount + i] = args.get(i);
-
-        return params;
-    }
-
-    public Object eval(String script, List<String> keys, List<String> args) {
-        return eval(script, keys.size(), getParams(keys, args));
-    }
-
-    public Object eval(String script) {
-        return eval(script, 0);
-    }
-
-    public Object evalsha(String script) {
-        return evalsha(script, 0);
-    }
-
-    private Object getEvalResult() {
-        return evalResult(client.getOne());
-    }
-
-    private Object evalResult(Object result) {
-        if (result instanceof byte[]) return SafeEncoder.encode((byte[]) result);
-
-        if (result instanceof List<?>) {
-            List<?> list = (List<?>) result;
-            List<Object> listResult = new ArrayList<Object>(list.size());
-            for (Object bin : list) {
-                listResult.add(evalResult(bin));
-            }
-
-            return listResult;
-        }
-
-        return result;
-    }
-
-    public Object evalsha(String sha1, List<String> keys, List<String> args) {
-        return evalsha(sha1, keys.size(), getParams(keys, args));
-    }
-
-    public Object evalsha(String sha1, int keyCount, String... params) {
-        checkIsInMulti();
-        client.evalsha(sha1, keyCount, params);
-        return getEvalResult();
-    }
-
-    public Boolean scriptExists(String sha1) {
-        String[] a = new String[1];
-        a[0] = sha1;
-        return scriptExists(a).get(0);
-    }
-
-    public List<Boolean> scriptExists(String... sha1) {
-        client.scriptExists(sha1);
-        List<Long> result = client.getIntegerMultiBulkReply();
-        List<Boolean> exists = new ArrayList<Boolean>();
-
-        for (Long value : result)
-            exists.add(value == 1);
-
-        return exists;
-    }
-
-    public String scriptLoad(String script) {
-        client.scriptLoad(script);
-        return client.getBulkReply();
-    }
-
-    public List<Slowlog> slowlogGet() {
-        client.slowlogGet();
-        return Slowlog.from(client.getObjectMultiBulkReply());
-    }
-
-    public List<Slowlog> slowlogGet(long entries) {
-        client.slowlogGet(entries);
-        return Slowlog.from(client.getObjectMultiBulkReply());
-    }
-
-    public Long objectRefcount(String string) {
-        client.objectRefcount(string);
-        return client.getIntegerReply();
-    }
-
-    public String objectEncoding(String string) {
-        client.objectEncoding(string);
-        return client.getBulkReply();
-    }
-
-    public Long objectIdletime(String string) {
-        client.objectIdletime(string);
-        return client.getIntegerReply();
-    }
-
-    public Long bitcount(final String key) {
-        client.bitcount(key);
-        return client.getIntegerReply();
-    }
-
-    public Long bitcount(final String key, long start, long end) {
-        client.bitcount(key, start, end);
-        return client.getIntegerReply();
-    }
-
-    public Long bitop(BitOP op, final String destKey, String... srcKeys) {
-        client.bitop(op, destKey, srcKeys);
-        return client.getIntegerReply();
-    }
-
-    @SuppressWarnings("rawtypes")
-    public List<Map<String, String>> sentinelMasters() {
-        client.sentinel(Protocol.SENTINEL_MASTERS);
-        final List<Object> reply = client.getObjectMultiBulkReply();
-
-        final List<Map<String, String>> masters = new ArrayList<Map<String, String>>();
-        for (Object obj : reply) {
-            masters.add(BuilderFactory.STRING_MAP.build((List) obj));
-        }
-        return masters;
-    }
-
-    public List<String> sentinelGetMasterAddrByName(String masterName) {
-        client.sentinel(Protocol.SENTINEL_GET_MASTER_ADDR_BY_NAME, masterName);
-        final List<Object> reply = client.getObjectMultiBulkReply();
-        return BuilderFactory.STRING_LIST.build(reply);
-    }
-
-    public Long sentinelReset(String pattern) {
-        client.sentinel(Protocol.SENTINEL_RESET, pattern);
-        return client.getIntegerReply();
-    }
-
-    @SuppressWarnings("rawtypes")
-    public List<Map<String, String>> sentinelSlaves(String masterName) {
-        client.sentinel(Protocol.SENTINEL_SLAVES, masterName);
-        final List<Object> reply = client.getObjectMultiBulkReply();
-
-        final List<Map<String, String>> slaves = new ArrayList<Map<String, String>>();
-        for (Object obj : reply) {
-            slaves.add(BuilderFactory.STRING_MAP.build((List) obj));
-        }
-        return slaves;
-    }
-
-    public String sentinelFailover(String masterName) {
-        client.sentinel(Protocol.SENTINEL_FAILOVER, masterName);
-        return client.getStatusCodeReply();
-    }
-
-    public String sentinelMonitor(String masterName, String ip, int port, int quorum) {
-        client.sentinel(Protocol.SENTINEL_MONITOR, masterName, ip, String.valueOf(port),
-                        String.valueOf(quorum));
-        return client.getStatusCodeReply();
-    }
-
-    public String sentinelRemove(String masterName) {
-        client.sentinel(Protocol.SENTINEL_REMOVE, masterName);
-        return client.getStatusCodeReply();
-    }
-
-    public String sentinelSet(String masterName, Map<String, String> parameterMap) {
-        int index = 0;
-        int paramsLength = parameterMap.size() * 2 + 2;
-        String[] params = new String[paramsLength];
-
-        params[index++] = Protocol.SENTINEL_SET;
-        params[index++] = masterName;
-        for (Entry<String, String> entry : parameterMap.entrySet()) {
-            params[index++] = entry.getKey();
-            params[index++] = entry.getValue();
-        }
-
-        client.sentinel(params);
-        return client.getStatusCodeReply();
-    }
-
-    public byte[] dump(final String key) {
-        checkIsInMulti();
-        client.dump(key);
-        return client.getBinaryBulkReply();
-    }
-
-    public String restore(final String key, final int ttl, final byte[] serializedValue) {
-        checkIsInMulti();
-        client.restore(key, ttl, serializedValue);
-        return client.getStatusCodeReply();
-    }
-
-    public Long pexpire(final String key, final long milliseconds) {
-        checkIsInMulti();
-        client.pexpire(key, milliseconds);
-        return client.getIntegerReply();
-    }
-
-    public Long pexpireAt(final String key, final long millisecondsTimestamp) {
-        checkIsInMulti();
-        client.pexpireAt(key, millisecondsTimestamp);
-        return client.getIntegerReply();
-    }
-
-    public Long pttl(final String key) {
-        checkIsInMulti();
-        client.pttl(key);
-        return client.getIntegerReply();
-    }
-
-    public String psetex(final String key, final long milliseconds, final String value) {
-        checkIsInMulti();
-        client.psetex(key, milliseconds, value);
-        return client.getStatusCodeReply();
-    }
-
-    public String set(final String key, final String value, final String nxxx) {
-        checkIsInMulti();
-        client.set(key, value, nxxx);
-        return client.getStatusCodeReply();
-    }
-
-    public String set(final String key, final String value, final String nxxx, final String expx,
-                      final int time) {
-        checkIsInMulti();
-        client.set(key, value, nxxx, expx, time);
-        return client.getStatusCodeReply();
-    }
-
-    public String clientKill(final String client) {
-        checkIsInMulti();
-        this.client.clientKill(client);
-        return this.client.getStatusCodeReply();
-    }
-
-    public String clientSetname(final String name) {
-        checkIsInMulti();
-        client.clientSetname(name);
-        return client.getStatusCodeReply();
-    }
-
-    public String migrate(final String host, final int port, final String key,
-                          final int destinationDb, final int timeout) {
-        checkIsInMulti();
-        client.migrate(host, port, key, destinationDb, timeout);
-        return client.getStatusCodeReply();
-    }
-
-    public ScanResult<String> scan(final String cursor) {
-        return scan(cursor, new ScanParams());
-    }
-
-    public ScanResult<String> scan(final String cursor, final ScanParams params) {
-        checkIsInMulti();
-        client.scan(cursor, params);
-        List<Object> result = client.getObjectMultiBulkReply();
-        String newcursor = new String((byte[]) result.get(0));
-        List<String> results = new ArrayList<String>();
-        List<byte[]> rawResults = (List<byte[]>) result.get(1);
-        for (byte[] bs : rawResults) {
-            results.add(SafeEncoder.encode(bs));
-        }
-        return new ScanResult<String>(newcursor, results);
-    }
-
-    public ScanResult<Map.Entry<String, String>> hscan(final String key, final String cursor) {
-        return hscan(key, cursor, new ScanParams());
-    }
-
-    public ScanResult<Map.Entry<String, String>> hscan(final String key, final String cursor,
-                                                       final ScanParams params) {
-        checkIsInMulti();
-        client.hscan(key, cursor, params);
-        List<Object> result = client.getObjectMultiBulkReply();
-        String newcursor = new String((byte[]) result.get(0));
-        List<Map.Entry<String, String>> results = new ArrayList<Map.Entry<String, String>>();
-        List<byte[]> rawResults = (List<byte[]>) result.get(1);
-        Iterator<byte[]> iterator = rawResults.iterator();
-        while (iterator.hasNext()) {
-            results.add(new AbstractMap.SimpleEntry<String, String>(SafeEncoder.encode(iterator.next()),
-                                                                    SafeEncoder.encode(iterator.next())));
-        }
-        return new ScanResult<Map.Entry<String, String>>(newcursor, results);
-    }
-
-    public ScanResult<String> sscan(final String key, final String cursor) {
-        return sscan(key, cursor, new ScanParams());
-    }
-
-    public ScanResult<String> sscan(final String key, final String cursor, final ScanParams params) {
-        checkIsInMulti();
-        client.sscan(key, cursor, params);
-        List<Object> result = client.getObjectMultiBulkReply();
-        String newcursor = new String((byte[]) result.get(0));
-        List<String> results = new ArrayList<String>();
-        List<byte[]> rawResults = (List<byte[]>) result.get(1);
-        for (byte[] bs : rawResults) {
-            results.add(SafeEncoder.encode(bs));
-        }
-        return new ScanResult<String>(newcursor, results);
-    }
-
-    
-
-    public ScanResult<Tuple> zscan(final String key, final String cursor, final ScanParams params) {
-        checkIsInMulti();
-        client.zscan(key, cursor, params);
-        List<Object> result = client.getObjectMultiBulkReply();
-        String newcursor = new String((byte[]) result.get(0));
-        List<Tuple> results = new ArrayList<Tuple>();
-        List<byte[]> rawResults = (List<byte[]>) result.get(1);
-        Iterator<byte[]> iterator = rawResults.iterator();
-        while (iterator.hasNext()) {
-            results.add(new Tuple(SafeEncoder.encode(iterator.next()), Double.valueOf(SafeEncoder
-                                                                                      .encode(iterator.next()))));
-        }
-        return new ScanResult<Tuple>(newcursor, results);
-    }
-
-    public String clusterNodes() {
-        checkIsInMulti();
-        client.clusterNodes();
-        return client.getBulkReply();
-    }
-
-    public String clusterMeet(final String ip, final int port) {
-        checkIsInMulti();
-        client.clusterMeet(ip, port);
-        return client.getStatusCodeReply();
-    }
-
-    public String clusterReset(final Reset resetType) {
-        checkIsInMulti();
-        client.clusterReset(resetType);
-        return client.getStatusCodeReply();
-    }
-
-    public String clusterAddSlots(final int... slots) {
-        checkIsInMulti();
-        client.clusterAddSlots(slots);
-        return client.getStatusCodeReply();
-    }
-
-    public String clusterDelSlots(final int... slots) {
-        checkIsInMulti();
-        client.clusterDelSlots(slots);
-        return client.getStatusCodeReply();
-    }
-
-    public String clusterInfo() {
-        checkIsInMulti();
-        client.clusterInfo();
-        return client.getStatusCodeReply();
-    }
-
-    public List<String> clusterGetKeysInSlot(final int slot, final int count) {
-        checkIsInMulti();
-        client.clusterGetKeysInSlot(slot, count);
-        return client.getMultiBulkReply();
-    }
-
-    public String clusterSetSlotNode(final int slot, final String nodeId) {
-        checkIsInMulti();
-        client.clusterSetSlotNode(slot, nodeId);
-        return client.getStatusCodeReply();
-    }
-
-    public String clusterSetSlotMigrating(final int slot, final String nodeId) {
-        checkIsInMulti();
-        client.clusterSetSlotMigrating(slot, nodeId);
-        return client.getStatusCodeReply();
-    }
-
-    public String clusterSetSlotImporting(final int slot, final String nodeId) {
-        checkIsInMulti();
-        client.clusterSetSlotImporting(slot, nodeId);
-        return client.getStatusCodeReply();
-    }
-
-    public String clusterSetSlotStable(final int slot) {
-        checkIsInMulti();
-        client.clusterSetSlotStable(slot);
-        return client.getStatusCodeReply();
-    }
-
-    public String clusterForget(final String nodeId) {
-        checkIsInMulti();
-        client.clusterForget(nodeId);
-        return client.getStatusCodeReply();
-    }
-
-    public String clusterFlushSlots() {
-        checkIsInMulti();
-        client.clusterFlushSlots();
-        return client.getStatusCodeReply();
-    }
-
-    public Long clusterKeySlot(final String key) {
-        checkIsInMulti();
-        client.clusterKeySlot(key);
-        return client.getIntegerReply();
-    }
-
-    public Long clusterCountKeysInSlot(final int slot) {
-        checkIsInMulti();
-        client.clusterCountKeysInSlot(slot);
-        return client.getIntegerReply();
-    }
-
-    public String clusterSaveConfig() {
-        checkIsInMulti();
-        client.clusterSaveConfig();
-        return client.getStatusCodeReply();
-    }
-
-    public String clusterReplicate(final String nodeId) {
-        checkIsInMulti();
-        client.clusterReplicate(nodeId);
-        return client.getStatusCodeReply();
-    }
-
-    public List<String> clusterSlaves(final String nodeId) {
-        checkIsInMulti();
-        client.clusterSlaves(nodeId);
-        return client.getMultiBulkReply();
-    }
-
-    public String clusterFailover() {
-        checkIsInMulti();
-        client.clusterFailover();
-        return client.getStatusCodeReply();
-    }
-
-    @Override
-    public List<Object> clusterSlots() {
-        checkIsInMulti();
-        client.clusterSlots();
-        return client.getObjectMultiBulkReply();
-    }
-
-    public String asking() {
-        checkIsInMulti();
-        client.asking();
-        return client.getStatusCodeReply();
-    }
-
-    public List<String> pubsubChannels(String pattern) {
-        checkIsInMulti();
-        client.pubsubChannels(pattern);
-        return client.getMultiBulkReply();
-    }
-
-    public Long pubsubNumPat() {
-        checkIsInMulti();
-        client.pubsubNumPat();
-        return client.getIntegerReply();
-    }
-
-    public Map<String, String> pubsubNumSub(String... channels) {
-        checkIsInMulti();
-        client.pubsubNumSub(channels);
-        return BuilderFactory.PUBSUB_NUMSUB_MAP.build(client.getBinaryMultiBulkReply());
-    }
-
-    @Override
-    public void close() {
-        if (dataSource != null) {
-            if (client.isBroken()) {
-                this.dataSource.returnBrokenResource(this);
-            } else {
-                this.dataSource.returnResource(this);
-            }
-        } else {
-            client.close();
-        }
-    }
-
-    public void setDataSource(JedisPoolAbstract jedisPool) {
-        this.dataSource = jedisPool;
-    }
-
-    public Long pfadd(final String key, final String... elements) {
-        checkIsInMulti();
-        client.pfadd(key, elements);
-        return client.getIntegerReply();
-    }
-
-    public long pfcount(final String key) {
-        checkIsInMulti();
-        client.pfcount(key);
-        return client.getIntegerReply();
-    }
-
-    @Override
-    public long pfcount(String... keys) {
-        checkIsInMulti();
-        client.pfcount(keys);
-        return client.getIntegerReply();
-    }
-
-    public String pfmerge(final String destkey, final String... sourcekeys) {
-        checkIsInMulti();
-        client.pfmerge(destkey, sourcekeys);
-        return client.getStatusCodeReply();
-    }
-
-    @Override
-    public List<String> blpop(int timeout, String key) {
-        return blpop(key, String.valueOf(timeout));
-    }
-
-    @Override
-    public List<String> brpop(int timeout, String key) {
-        return brpop(key, String.valueOf(timeout));
-    }
-    */
 }
