+++ redis-persistence/src/main/java/com/netflix/conductor/jedis/RedisClusterJedisProvider.java
@@ -1,18 +1,25 @@
+/*
+ * Copyright 2019 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 package com.netflix.conductor.jedis;
 
 import com.netflix.dyno.connectionpool.Host;
 import com.netflix.dyno.connectionpool.HostSupplier;
-
-import org.apache.commons.pool2.impl.GenericObjectPoolConfig;
-
 import java.util.ArrayList;
-
 import javax.inject.Inject;
 import javax.inject.Provider;
-
-import redis.clients.jedis.HostAndPort;
-import redis.clients.jedis.JedisCluster;
-import redis.clients.jedis.JedisCommands;
+import org.apache.commons.pool2.impl.GenericObjectPoolConfig;
+import redis.clients.jedis.JedisPool;
+import redis.clients.jedis.commands.JedisCommands;
 
 public class RedisClusterJedisProvider implements Provider<JedisCommands> {
 
@@ -26,10 +33,12 @@ public class RedisClusterJedisProvider implements Provider<JedisCommands> {
     @Override
     public JedisCommands get() {
         // FIXME This doesn't seem very safe, but is how it was in the code this was moved from.
-        Host host = new ArrayList<Host>(hostSupplier.getHosts()).get(0);
+        Host host = new ArrayList<>(hostSupplier.getHosts()).get(0);
         GenericObjectPoolConfig poolConfig = new GenericObjectPoolConfig();
         poolConfig.setMinIdle(5);
         poolConfig.setMaxTotal(1000);
-        return new JedisCluster(new HostAndPort(host.getHostName(), host.getPort()), poolConfig);
+
+        JedisPool jedisPool = new JedisPool(poolConfig, host.getHostName(), host.getPort());
+        return new JedisCluster(jedisPool);
     }
 }
