+++ redis-persistence/src/main/java/com/netflix/conductor/jedis/RedisSentinelJedisProvider.java
@@ -1,22 +1,33 @@
+/*
+ * Copyright 2019 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 package com.netflix.conductor.jedis;
 
 import com.netflix.conductor.dyno.DynomiteConfiguration;
 import com.netflix.dyno.connectionpool.Host;
 import com.netflix.dyno.connectionpool.HostSupplier;
+import java.util.HashSet;
+import java.util.Set;
+import javax.inject.Inject;
+import javax.inject.Provider;
 import org.apache.commons.pool2.impl.GenericObjectPoolConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import redis.clients.jedis.JedisCommands;
 import redis.clients.jedis.JedisSentinelPool;
-
-import javax.inject.Inject;
-import javax.inject.Provider;
-import java.util.HashSet;
-import java.util.Set;
+import redis.clients.jedis.commands.JedisCommands;
 
 public class RedisSentinelJedisProvider implements Provider<JedisCommands> {
     private static Logger logger = LoggerFactory.getLogger(RedisSentinelJedisProvider.class);
-    private final JedisSentinelPool jedisPool;
+    private final JedisSentinelPool jedisSentinelPool;
 
     @Inject
     public RedisSentinelJedisProvider(HostSupplier hostSupplier, DynomiteConfiguration configuration) {
@@ -32,11 +43,11 @@ public class RedisSentinelJedisProvider implements Provider<JedisCommands> {
             sentinels.add(host.getHostName() + ":" + host.getPort());
         }
 
-        jedisPool = new JedisSentinelPool(configuration.getClusterName(), sentinels, poolConfig);
+        jedisSentinelPool = new JedisSentinelPool(configuration.getClusterName(), sentinels, poolConfig);
     }
 
     @Override
     public JedisCommands get() {
-        return new JedisClusterSentinel(jedisPool);
+        return new JedisCluster(jedisSentinelPool);
     }
 }
