+++ redis-persistence/src/test/java/com/netflix/conductor/dao/dynomite/DynoQueueDAOTest.java
@@ -1,5 +1,5 @@
-/**
- * Copyright 2016 Netflix, Inc.
+/*
+ * Copyright 2019 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,29 +15,25 @@
  */
 package com.netflix.conductor.dao.dynomite;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
 import com.netflix.conductor.config.TestConfiguration;
-import com.netflix.conductor.common.utils.JsonMapperProvider;
 import com.netflix.conductor.dao.QueueDAO;
 import com.netflix.conductor.dao.dynomite.queue.DynoQueueDAO;
 import com.netflix.conductor.dao.redis.JedisMock;
+import com.netflix.dyno.connectionpool.Host;
 import com.netflix.dyno.queues.ShardSupplier;
-
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-import redis.clients.jedis.JedisCommands;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import redis.clients.jedis.commands.JedisCommands;
 
 /**
  * 
@@ -46,22 +42,25 @@ import static org.junit.Assert.assertNotNull;
  */
 public class DynoQueueDAOTest {
 
-	private QueueDAO dao;
-
-	private static ObjectMapper om = new JsonMapperProvider().get();
+	private QueueDAO queueDAO;
 
 	@Before
-	public void init() throws Exception {
+	public void init() {
 		JedisCommands jedisMock = new JedisMock();
-		dao = new DynoQueueDAO(jedisMock, jedisMock, new ShardSupplier() {
+		queueDAO = new DynoQueueDAO(jedisMock, jedisMock, new ShardSupplier() {
 
 			@Override
 			public Set<String> getQueueShards() {
-				return Arrays.asList("a").stream().collect(Collectors.toSet());
+				return new HashSet<>(Collections.singletonList("a"));
 			}
 
 			@Override
 			public String getCurrentShard() {
+				return "a";
+			}
+
+			@Override
+			public String getShardForHost(Host host) {
 				return "a";
 			}
 		}, new TestConfiguration());
@@ -77,67 +76,65 @@ public class DynoQueueDAOTest {
 		
 		for(int i = 0; i < 10; i++) {
 			String messageId = "msg" + i;
-			dao.push(queueName, messageId, offsetTimeInSecond);
+			queueDAO.push(queueName, messageId, offsetTimeInSecond);
 		}
-		int size = dao.getSize(queueName);
+		int size = queueDAO.getSize(queueName);
 		assertEquals(10, size);
-		Map<String, Long> details = dao.queuesDetail();
+		Map<String, Long> details = queueDAO.queuesDetail();
 		assertEquals(1, details.size());
 		assertEquals(10L, details.get(queueName).longValue());
 		
 		
 		for(int i = 0; i < 10; i++) {
 			String messageId = "msg" + i;
-			dao.pushIfNotExists(queueName, messageId, offsetTimeInSecond);
+			queueDAO.pushIfNotExists(queueName, messageId, offsetTimeInSecond);
 		}
 		
-		List<String> popped = dao.pop(queueName, 10, 100);
+		List<String> popped = queueDAO.pop(queueName, 10, 100);
 		assertNotNull(popped);
 		assertEquals(10, popped.size());
 		
-		Map<String, Map<String, Map<String, Long>>> verbose = dao.queuesDetailVerbose();
+		Map<String, Map<String, Map<String, Long>>> verbose = queueDAO.queuesDetailVerbose();
 		assertEquals(1, verbose.size());
 		long shardSize = verbose.get(queueName).get("a").get("size");
 		long unackedSize = verbose.get(queueName).get("a").get("uacked");
 		assertEquals(0, shardSize);
 		assertEquals(10, unackedSize);
 		
-		popped.forEach(messageId -> dao.ack(queueName, messageId));
+		popped.forEach(messageId -> queueDAO.ack(queueName, messageId));
 		
-		verbose = dao.queuesDetailVerbose();
+		verbose = queueDAO.queuesDetailVerbose();
 		assertEquals(1, verbose.size());
 		shardSize = verbose.get(queueName).get("a").get("size");
 		unackedSize = verbose.get(queueName).get("a").get("uacked");
 		assertEquals(0, shardSize);
 		assertEquals(0, unackedSize);
 		
-		popped = dao.pop(queueName, 10, 100);
+		popped = queueDAO.pop(queueName, 10, 100);
 		assertNotNull(popped);
 		assertEquals(0, popped.size());
 		
 		for(int i = 0; i < 10; i++) {
 			String messageId = "msg" + i;
-			dao.pushIfNotExists(queueName, messageId, offsetTimeInSecond);
+			queueDAO.pushIfNotExists(queueName, messageId, offsetTimeInSecond);
 		}
-		size = dao.getSize(queueName);
+		size = queueDAO.getSize(queueName);
 		assertEquals(10, size);
 		
 		for(int i = 0; i < 10; i++) {
 			String messageId = "msg" + i;
-			dao.remove(queueName, messageId);
+			queueDAO.remove(queueName, messageId);
 		}
 		
-		size = dao.getSize(queueName);
+		size = queueDAO.getSize(queueName);
 		assertEquals(0, size);
 		
 		for(int i = 0; i < 10; i++) {
 			String messageId = "msg" + i;
-			dao.pushIfNotExists(queueName, messageId, offsetTimeInSecond);
+			queueDAO.pushIfNotExists(queueName, messageId, offsetTimeInSecond);
 		}
-		dao.flush(queueName);
-		size = dao.getSize(queueName);
+		queueDAO.flush(queueName);
+		size = queueDAO.getSize(queueName);
 		assertEquals(0, size);
-		
 	}
-
 }
