+++ versionsOfDependencies.gradle
@@ -10,7 +10,7 @@ ext {
     revCassandraUnit = '3.5.0.1'
     revCommonsLang3 = '3.0'
     revCommonsIo = '2.4'
-    revDynoQueues = '2.0.0-rc8'
+    revDynoQueues = '2.0.13'
     revElasticSearch5 = '5.6.8'
     revElasticSearch5Client = '5.6.8'
     revElasticSearch6 = '6.5.1'
@@ -30,28 +30,28 @@ ext {
     revJavaxInject = '1'
     revJacksonCore = '2.7.5'
     revJacksonDatabind = '2.7.5'
-    revJedis = '2.9.0'
+    revJedis = '3.0.+'
     revJerseyBundle = '1.19.1'
     revJerseyClient = '1.19.4'
     revJerseyGuice = '1.19.4'
+    revJetteyServer = '9.3.9.v20160517'
+    revJettyServlet = '9.3.9.v20160517'
     revJUnit = '4.12'
     revJsr311Api = '1.1.1'
     revJq = '0.0.8'
     revLog4jApi = '2.9.1'
     revLog4jCore = '2.9.1'
     revMariaDB4j = '2.2.3'
-    revRxJava = '1.2.2'
     revMockito = '1.10.19'
     revMySqlConnector = '8.0.11'
     revNatsStreaming = '0.5.0'
-    revJetteyServer = '9.3.9.v20160517'
-    revJettyServlet = '9.3.9.v20160517'
     revOauthClient = '1.19.4'
     revOauthSignature = '1.19.4'
     revProtoBuf = '3.5.1'
     revProtogenAnnotations = '1.0.0'
     revProtogenCodegen = '1.4.0'
     revRarefiedRedis = '0.0.17'
+    revRxJava = '1.2.2'
     revServo = '0.12.17'
     revServletApi = '3.1.0'
     revSpotifyCompletableFutures = '0.3.1'
